requirements.txt
pip~=25.1.1
python-dotenv
python-telegram-bot
aiofiles